# ENH, DOC: Improve foreign object interoperability