As discussed in GH-28024, the logic with which functions in `numpy/_core/fromnumeric.py` navigate methods in foreign objects is not clearly documented.

The major problem arises for object that *do not* implement the [`__array_function__`](https://numpy.org/devdocs/user/basics.interoperability.html#the-array-function-protocol)  protocol: the fact that say `np.mean(obj)` will call `obj.mean(axis=None, dtype=None, out=None)` even *before* trying `obj.__array__` is not documented.

In this PR:
- [x] add checks in numpy code to avoid calling non callable attributes
- [ ] add note in numpy docs to warn developers about order in which methods are tested if `__array_function__` not present
- [ ] gently nudge developers to adopt newer `__array_function__` protocol