@miccoli 2025-01-04
> This works, but I somewhat dislike that it slows down the most common case, where it does call the attribute correctly. See in-line comments for a possible solution.

These are python functions, not ufuncs, so I think that this concern is misplaced: I tried to favor clarity of intent over speed, while keeping the changes to a minimum.

If necessary I can setup a proper benchmark, however just for curiosity I wrote this simple `timeit` script:

<details>

```python
import timeit

import numpy as np

p0 = """
import numpy as np
a = np.arange(1_000, dtype=np.float64).reshape(20, -1)
"""

bench = ["np.sum(a)", "a.sum()", "np.cumsum(a)", "a.cumsum()"]
n = 100_000
r = 10
scale = 1e6


print(np.version.version)
print(p0)
print(f"n = {n:_d}, r = {r:d}")

for s in bench:
    print(f"{s:12s}", end="")
    t = timeit.Timer(stmt=s, setup=p0)
    t1 = np.array(t.repeat(repeat=r, number=n)) / n * scale
    print(f"  min = {t1.min():.3f} µs, mean = ({t1.mean():.3f} ± {t1.std():.3f}) µs ")
```

</details>

The results on my (old) laptop are:

#### baseline

```
2.3.0.dev0+git20250102.12364e3

import numpy as np
a = np.arange(1_000, dtype=np.float64).reshape(20, -1)

n = 100_000, r = 10
np.sum(a)     min = 2.847 µs, mean = (2.919 ± 0.042) µs
a.sum()       min = 1.483 µs, mean = (1.533 ± 0.044) µs
np.cumsum(a)  min = 4.017 µs, mean = (4.068 ± 0.050) µs
a.cumsum()    min = 3.021 µs, mean = (3.389 ± 0.973) µs
```

#### after this PR

```
2.3.0.dev0+git20250103.b9a8d9e

import numpy as np
a = np.arange(1_000, dtype=np.float64).reshape(20, -1)

n = 100_000, r = 10
np.sum(a)     min = 3.153 µs, mean = (3.328 ± 0.098) µs
a.sum()       min = 1.743 µs, mean = (1.778 ± 0.030) µs
np.cumsum(a)  min = 4.120 µs, mean = (4.200 ± 0.057) µs
a.cumsum()    min = 3.096 µs, mean = (3.134 ± 0.031) µs
```

My comment here is that this PR for sure add a small overhead, but this is so small (< 0.5 µs per call) that it is very hard to measure it properly, at least on a laptop.

Here is a second run of the very same benchmark, after the system awaked from sleep: you can see that now the code after the PR is even faster than the baseline. This clearly shows that on a multi-user system, the operating conditions have a big influence on the figures.
<details>

```
2.3.0.dev0+git20250103.b9a8d9e

import numpy as np
a = np.arange(1_000, dtype=np.float64).reshape(20, -1)

n = 100_000, r = 10
np.sum(a)     min = 2.700 µs, mean = (2.751 ± 0.038) µs
a.sum()       min = 1.489 µs, mean = (1.538 ± 0.029) µs
np.cumsum(a)  min = 3.997 µs, mean = (4.106 ± 0.073) µs
a.cumsum()    min = 2.958 µs, mean = (3.010 ± 0.032) µs
```
</details>


### final considerations

Properly benchmarking this changes would be very a complex task (meaningful bechmarking is an art in itself...) but I do not think that it makes much sense: if in a given “numpy intensive” program the bottleneck are the overheads due to python function calls, then probably the program should be refactored, instead of over-optimizing the numpy code-base.