@mhvk 2025-01-04
But why increase the overhead at all if it is not needed? This is code that is used a *lot*, so speed does matter. (Also, you use fairly large arrays; for small ones it would be worse.)