@miccoli 2025-01-04
> But why increase the overhead at all if it is not needed? This is code that is used a _lot_, so speed does matter. (Also, you use fairly large arrays; for small ones it would be worse.)

My point is that nobody will notice. Please have a look at the above timings: actually only `np.cumsum(a)` goes through the new branches (via `_wrapfunc`). All the other function calls should not be touched by the changes. The small timing differences are close to noise.