@mhvk 2025-01-04
Please believe me: people *will* notice an extra 0.3 us -- that's a 17% increase for `a.sum()` (for a relatively large array; the fractional increase will be worse for smaller ones).  There's simply no reason to pay that price for a work-around for a problem that affects very few people, especially if for those performance doesn't matter at all since an error is going to get raised anyway.