@miccoli 2025-01-04
> Please believe me: people _will_ notice an extra 0.3 us -- that's a 17% increase for `a.sum()` (for a relatively large array; the fractional increase will be worse for smaller ones).

`a.sum()` is not affected by the changes, it follows a different path, not affected by this PR. I'm timing it because `np.sum(a)` will dispatch to `a.sum()`, to have a reference.

Other timings, with `n = 1_000_000` in which the differences become smaller:
<details>

```
2.3.0.dev0+git20250102.12364e3

# import numpy as np
a = np.arange(1_000, dtype=np.float64).reshape(20, -1)

n = 1_000_000, r = 10
np.sum(a)     min = 2.956 µs, mean = (3.054 ± 0.095) µs
a.sum()       min = 1.569 µs, mean = (1.663 ± 0.151) µs
np.cumsum(a)  min = 4.106 µs, mean = (4.230 ± 0.148) µs
a.cumsum()    min = 3.079 µs, mean = (3.135 ± 0.103) µs
```

```
2.3.0.dev0+git20250103.b9a8d9e

# import numpy as np
a = np.arange(1_000, dtype=np.float64).reshape(20, -1)

n = 1_000_000, r = 10
np.sum(a)     min = 3.053 µs, mean = (3.188 ± 0.123) µs
a.sum()       min = 1.584 µs, mean = (1.611 ± 0.025) µs
np.cumsum(a)  min = 4.134 µs, mean = (4.237 ± 0.117) µs
a.cumsum()    min = 3.093 µs, mean = (3.236 ± 0.157) µs
```

`timeit` seems not to be adequate here to catch the overhead... As said a proper benchmark, encompassing a variety of matrix sizes and different functions should be set up. I'm confident that the measured overhead should be much smaller than the 0.5 µs that I mentioned above as a very loose upper limit ...

</details>

Sorry but I remain of my opinion: you will notice the overhead only if you call `np.cumsum(a)` in a tight loop with smallish `a` (and for me 20×50 is smallish), but in such cases a code refactor for vectorizing the loop, or rewriting in a compiled language would be the optimal solution.

BTW thanks for the discussion.